move getworkerusecase tests into workerusecase tests
